CPU Optimization
Optimisation with the infinite while loop by sleeping the thread for 2 seconds when the user is not logged in. Meaning, they are not within close enough time to the reservation time to be running at full speed. Got the CPU usage down from around ~14% to ~3%. (that's ~3% while the thread is not sleeping)